// apps/landing/src/types/landingtypes.ts
export interface NavMenuItem {
  name: string;
  link: string;
}

export interface BrandLogoItem {
  src: string;
  alt: string;
}

export interface SolutionItem {
  icon: string;
  body: string;
}

export interface OperationItem {
  title: string;
  body: string;
  label: string;
  ratio: string;
}

export interface FaqItem {
  q: string;
  a: string;
}